refactor: submit a request to the server in a separate method
