+++ src/components/SearchPanel/CardList/CardList.tsx
@@ -5,13 +5,16 @@ import { upperFirstLetter } from '../../../utils/helpers';
 
 class CardList extends React.Component<CardListProps> {
   public render() {
+    const { results, error } = this.props;
+
     return (
       <>
         <table>
           <thead>
             <tr>
-              {this.props.results.length > 0 &&
-                Object.keys(this.props.results[0]).map((key) => {
+              {Array.isArray(results) &&
+                results.length > 0 &&
+                Object.keys(results[0]).map((key) => {
                   if (key !== 'id') {
                     return <th key={key}>{upperFirstLetter(key)}</th>;
                   }
@@ -20,13 +23,11 @@ class CardList extends React.Component<CardListProps> {
             </tr>
           </thead>
           <tbody>
-            {this.props.results.length > 0 ? (
-              this.props.results.map((item) => (
-                <Card key={item.id} data={item} />
-              ))
+            {Array.isArray(results) && results.length > 0 ? (
+              results.map((item) => <Card key={item.id} data={item} />)
             ) : (
               <tr>
-                <td>{this.props.error ?? 'Not results found'}</td>
+                <td>{error ?? 'No results found'}</td>
               </tr>
             )}
           </tbody>
